fix(setup.py): change how we handle install command
Due to the way setuptools.command.install.run handles detects
install command, sub classing Install breaks that login and thus
calling the `run` from super class results in a broken setup
process.
The setup process would finish without any errors but it won't
install the requirements specified in the setup.py which is not
expected. CI doesn't notice it because it uses `tox` to run tests
which installs dependacies itself.
In this change we try to reiterate the logic used in the super
class's defination of run and do the right thing for install
command and fallback to super class implmentation for the other
cases.

Fixes: #997