feat(semantics): emit type casts after doing arg type checks
In this commit we start emitting type casts for the arguments in
function calls, mutations or services.
We were already doing the checks for arg compatibility and had
implicit cast method ready as well. In this commit we just refactor
the implicit_cast method a bit to extract out a type_cast_entity
function which works on single entity and add type casts to its AST
by adding as operator to its tree.
As a result, now calling a function which accpets a float parameter
with an int parameter will specifically emit a type cast
(as operator) in the AST.

Closes: #993.