feat(parser): support expressions in list and object index
In this commit we start supporting expressions in the list or
object index. This means following syntax is now supported:
Examples:
		a[2 + 2]
		a["foo" + "bar"]

We do this by modifying grammar to accept expressions as list or
an object index and work on path resolution code path to interpret
these expressions and spit out correct resolved types.

Closes: #559.