fix(ExpressionResolver): make type_to_tree() handle `any` type
In this commit we fix the issue which arose due to the fact that
type_to_tree() didn't handle `any` type. This results in assertion
errors since give a type Map[string, any] this function can call
itself with type `any` and end up with errors.
In general as well, this function should have been able to handle
`any` type which we now add to it.

Fixes: #1072.