chore(Features): add service typing features flag
This is a temporary flag we are going to use to ease in the
service typing work into the compiler. By means of this flag
we will be disabling service typing for tests so that we can
gradually migrate them without creating huge commits.

This will go away after the tests have been migrated and the
feature is stable.