refactor(Functions): extract logic to type check and emit type cast
In this we extract out a helpers.py where we keep the logic for
doing the type check and also emit a implicit type cast for args

This will be usefull in adding the same kinda type check in the args
for services as well. We add that in upcoming commit.