fix(ExpressionResolver): change parameter name to avoid conflict
`name` variable was being used inside the build_arguments() body
inside the for loop for a different purpose then what it was being
passed along for. As a result there was a conflict of interest
between the two usage of same variable. We change parameter variable
name to fix this.