chore(Types): ban index access to object
We do this since for objects, we are performing checks on property
existance and to support index access we would have had to evaluate
the index to perform the property existance and type check. This
would essentially be possible in the runtime and since storyscript
is statically typed we want this kinda checks to happen at compile
time. Therefore we won't support index access to objects.