feat(ObjectType): ban cast to object and return an object
We aren't sure about what we want ObjectType to be. There is
confusion with Maps if we start calling {} to be an object?

Hence for the moment we decided to stop allowing objects
being returned from functions or other datatypes being
casted to an object. Well only datatype that can be casted
to an object is `any` so we ban that as well.