feat(TypeResolver): check existance of event based action output in when
We make sure that:

* The event based action output in when block exists and
is the output of parent service block only in case when block is
nested inside a service block.
Eg.
```
a = 1
http server
    when a listen path: "/" as r
        r write content: "foobar"
```
Above errors out because `a` in when block isn't output of parent
service block.

```
http server
    when client listen path: "/" as r
        r write content: "foobar"
```
Above errors out because `client` in when block does not exist.

* For syntax like below:
```
a = http server
when a listen path: "/" as r
    r write content: "foobar"
```

we make sure that `a` exists and is of type `object` and is wrapping
the output of the `Service`.