chore(PathResolver): pass more exact tree node to range deprecate
This totally fixes the error message arrow highlight for cases
with range_start_end but just improves it for the range_start
and range_end by not including the path.
This could be improved in future even more but right now there
is no solution available which very quickly fixes this.
This is because information about the `:` token is lost during
parsing.
We have open issue #1248 for retaining such info about `:` and
will improve these errors further with the said issue.