chore(Types): remove special case for AnyType
We can remove this since we migrated AnyType to behave like
unknown and hence any can no longer be assigned to anything
but any. Therefore AnyType.can_be_assigned will always return
None which the special case was added to do before the migration
of AnyType to behave like unknown.